RAG/requirements.txt
requests
python-dotenv
qdrant-client
elasticsearch>=8.14.0,<9
